pyproject.toml
[tool.poetry]
name = "kafka-flink-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
confluent-kafka = "^2.3.0"
requests = "^2.31.0"
fastavro = "^1.9.0"
python-dotenv = "^1.0.0"
avro = "^1.11.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
